requirements.txt
Flask == 1.1.2
Flask-SQLAlchemy == 2.4.4
Flask-Login == 0.5.0
click == 7.1.2
psycopg2-binary == 2.8.6
cassandra-driver == 3.24.0
itsdangerous == 1.1.0
pytest == 6.1.2
pillow == 8.0.1
cryptography == 3.3
gunicorn == 20.0.4
simplejson == 3.17.2
pandas == 1.1.5
